+++ Kernel/Net/NetworkAdapter.h
@@ -81,8 +81,8 @@ protected:
     NetworkAdapter();
     void set_interface_name(const StringView& basename);
     void set_mac_address(const MACAddress& mac_address) { m_mac_address = mac_address; }
-    virtual void send_raw(const u8*, int) = 0;
-    void did_receive(const u8*, int);
+    virtual void send_raw(const u8*, size_t) = 0;
+    void did_receive(const u8*, size_t);
 
 private:
     MACAddress m_mac_address;
